requirements.txt
googlesearch-python==1.2.5
pandas==2.2.2